add terms conditions before google analytcis

+++ app/src/main/java/com/example/mooka_umkm/network/model/Product.kt
@@ -15,7 +15,7 @@ data class Product(
     val umkm_id: Int,
     val updated_at: String,
     val description: String,
-    val berat: Int,
+    val berat: Double,
     val umkm: UMKM
 ) : DataResponse<Product> {
     override fun retrieveData(): Product = this
